Remove master_aoame-fuseki workflow and refactor Azure deployment to Docker
The commit removes the master_aoame-fuseki.yml GitHub workflow, which was meant for deploying a Java application to Azure Web App. It also refactors the deployment procedure in azure-webapps-java-jar.yml to use Docker instead. The refactored workflow file has been renamed to azure-webapps-docker.yml, streamlining the deployment process and providing more flexibility in terms of the environment.